feat(mda): move `mda.py` to new `trajectory` folder
